Split Qwen35 validation helpers from runtime
